refactored all non screen related files out of the screens package
