+++ mazes.csv
@@ -8,3 +8,6 @@ id,name,author,creation_time,rows,cols,state,startPosition
 6,mazeName,authorName,2022-12-07, 17, 25, #########################:#S......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#......................E#:#########################:, [1 1]
 7,mazeName,authorName,2022-12-08, 17, 25, #########################:#S......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#......................E#:#########################:, [1 1]
 8,mazeName,authorName,2022-12-08, 17, 25, #########################:#S......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#......................E#:#########################:, [1 1]
+9,mazeName,authorName,2022-12-08, 17, 25, #########################:#S......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#......................E#:#########################:, [1 1]
+10,mazeName,authorName,2022-12-08, 17, 25, #########################:#S......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#......................E#:#########################:, [1 1]
+11,mazeName,authorName,2022-12-08, 17, 25, #########################:#S......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#.......................#:#......................E#:#########################:, [1 1]
