+++ src/main/java/singletons/MazeInformation.java
@@ -1,4 +1,4 @@
-package screens;
+package singletons;
 
 
 import java.time.LocalDate;
