+++ src/main/java/singletons/UserSingleton.java
@@ -1,4 +1,4 @@
-package screens;
+package singletons;
 
 public class UserSingleton {
