+++ users.csv
@@ -1,7 +1,7 @@
 username,user_type,password,creation_time,mazes_played
 ween,Designer,asdf,2022-12-07 10:42,[]
 eshwara,Designer,eshwara,2022-12-07 17:29,[]
-eshwar,Player,eshwar,2022-12-07 17:28,3:
+eshwar,Player,eshwar,2022-12-07 17:28,3:2:4:7:9:
 oscar,Player,oscar,2022-12-08 10:54,[]
 edwin,Designer,1234,2022-12-06 22:22,[]
 mow,Player,123,2022-12-05 21:49,1:
